syslog: Cache connections
ecs-logs was opening a new syslog TCP connection for every message
batch.  This had many undesirable downstream resource utilization
impacts (among them, port exhaustion on our SOCKS proxies) and is
arguably unnecessary anyway.   Also, opening a new connection
each time causes the TCP session (either from the SOCKS proxy, or
directly from the client) not to be able to take advantage of
window scaling.

With this change, we now keep each syslog connection in a cache (keyed
by destination) and reuse it if we can.